// src/app/module/auth/auth.service.ts
import { UserStatus } from "../../../generated/prisma/enums";
import { auth } from "../../lib/auth";
import status from "http-status";
import AppError from "../../errorHelpers/AppError";
import { tokenUtils } from "../../utils/token";
import { prisma } from "../../lib/prisma";
import { jwtUtils } from "../../utils/jwt";
import { envVars } from "../../config/env";
import { JwtPayload } from "jsonwebtoken";
import { IChangePasswordPayload } from "./auth.interface";
import { IEmailOtpPayload } from "./auth.controller";


interface ILoginUserPayload {
    email: string;
    password: string;
}


// const loginUser = async (payload: ILoginUserPayload) => {
//     const { email, password } = payload;
//     const data = await auth.api.signInEmail({
//         body: {
//             email,
//             password,
//         }
//     })

//    if (data.user.status === UserStatus.BLOCKED) {
//         throw new AppError(status.FORBIDDEN, "User is blocked");
//     }

//     if (data.user.isDeleted || data.user.status === UserStatus.DELETED) {
//         throw new AppError(status.NOT_FOUND, "User is deleted");
//     }

//     const accessToken = tokenUtils.getAccessToken({
//         userId: data.user.id,
//         role: data.user.role,
//         name: data.user.name,
//         email: data.user.email,
//         status: data.user.status,
//         isDeleted: data.user.isDeleted,
//         emailVerified: data.user.emailVerified,
//     });

//     const refreshToken = tokenUtils.getRefreshToken({
//         userId: data.user.id,
//         role: data.user.role,
//         name: data.user.name,
//         email: data.user.email,
//         status: data.user.status,
//         isDeleted: data.user.isDeleted,
//         emailVerified: data.user.emailVerified,
//     });

//     return {
//         ...data,
//         accessToken,
//         refreshToken,
//     };


// }

const loginUser = async (payload: ILoginUserPayload) => {
    const { email, password } = payload;

    const data = await auth.api.signInEmail({
        body: {
            email,
            password,
        }
    })

    if (data.user.status === UserStatus.BLOCKED) {
        throw new AppError(status.FORBIDDEN, "User is blocked");
    }

    if (data.user.isDeleted || data.user.status === UserStatus.DELETED) {
        throw new AppError(status.NOT_FOUND, "User is deleted");
    }

    const accessToken = tokenUtils.getAccessToken({
        userId: data.user.id,
        role: data.user.role,
        name: data.user.name,
        email: data.user.email,
        status: data.user.status,
        isDeleted: data.user.isDeleted,
        emailVerified: data.user.emailVerified,
    });

    const refreshToken = tokenUtils.getRefreshToken({
        userId: data.user.id,
        role: data.user.role,
        name: data.user.name,
        email: data.user.email,
        status: data.user.status,
        isDeleted: data.user.isDeleted,
        emailVerified: data.user.emailVerified,
    });

    return {
        ...data,
        accessToken,
        refreshToken,
    };

}

const getNewToken = async (refreshToken : string, sessionToken : string) => {

    const isSessionTokenExists = await prisma.session.findUnique({
        where : {
            token : sessionToken,
        },
        include : {
            user : true,
        }
    })

    if(!isSessionTokenExists){
        throw new AppError(status.UNAUTHORIZED, "Invalid session token");
    }

    const verifiedRefreshToken = jwtUtils.verifyToken(refreshToken, envVars.REFRESH_TOKEN_SECRET)


    if(!verifiedRefreshToken.success && verifiedRefreshToken.error){
        throw new AppError(status.UNAUTHORIZED, "Invalid refresh token");
    }

    const data = verifiedRefreshToken.data as JwtPayload;

    const newAccessToken = tokenUtils.getAccessToken({
        userId: data.userId,
        role: data.role,
        name: data.name,
        email: data.email,
        status: data.status,
        isDeleted: data.isDeleted,
        emailVerified: data.emailVerified,
    });

    const newRefreshToken = tokenUtils.getRefreshToken({
        userId: data.userId,
        role: data.role,
        name: data.name,
        email: data.email,
        status: data.status,
        isDeleted: data.isDeleted,
        emailVerified: data.emailVerified,
    });

    const {token} = await prisma.session.update({
        where : {
            token : sessionToken
        },
        data : {
            token : sessionToken,
            expiresAt: new Date(Date.now() + 60 * 60 * 60 * 24 * 1000),
            updatedAt: new Date(),
        }
    })

    return {
        accessToken : newAccessToken,
        refreshToken : newRefreshToken,
        sessionToken : token,
    }

}

const changePassword = async (
  payload: IChangePasswordPayload,
  sessionToken: string
) => {


  if (!sessionToken) {
    console.log("❌ No session token found");
    throw new AppError(status.UNAUTHORIZED, "Session token missing");
  }

  const { currentPassword, newPassword } = payload;

  console.log("🟡 Payload received:", {
    currentPassword,
    newPassword,
  });

  try {
    const session = await auth.api.getSession({
      headers: new Headers({
        Authorization: `Bearer ${sessionToken}`
      })
    });

     if(!session){
        throw new AppError(status.UNAUTHORIZED, "Invalid session token");
    }

    const result = await auth.api.changePassword({
      body: {
        currentPassword,
        newPassword,
        revokeOtherSessions: false,
      },
      headers: {
        cookie: `better-auth.session_token=${sessionToken}`,
      },
    });

        if (session?.user?.needPasswordChange) {
        await prisma.user.update({
            where: { id: session.user.id },
            data: { needPasswordChange: false }
        })
    }

     const accessToken = tokenUtils.getAccessToken({
        userId: session.user.id,
        role: session.user.role,
        name: session.user.name,
        email: session.user.email,
        status: session.user.status,
        isDeleted: session.user.isDeleted,
        emailVerified: session.user.emailVerified,
    });

    const refreshToken = tokenUtils.getRefreshToken({
        userId: session.user.id,
        role: session.user.role,
        name: session.user.name,
        email: session.user.email,
        status: session.user.status,
        isDeleted: session.user.isDeleted,
        emailVerified: session.user.emailVerified,
    });


    return {
        ...result,
        accessToken,
        refreshToken,
    };

  } catch (error: any) {
    throw new AppError(
      status.UNAUTHORIZED,
      error?.message || "Password change failed"
    );
  }
};


const logoutUser = async (sessionToken : string) => {
    const result = await auth.api.signOut({
        headers : new Headers({
            Authorization : `Bearer ${sessionToken}`
        })
    })

    return result;
}

const verifyEmail = async (payload: IEmailOtpPayload) => {
  const { email, otp } = payload;

  const result = await auth.api.verifyEmailOTP({
    body: {
      email,
      otp,
    },
  });

  if (result?.status && result.user) {
    await prisma.user.update({
      where: { email },
      data: {
        emailVerified: true,
      },
    });
  }

  return result;
};

const forgetPassword = async (email : string) => {
    const isUserExist = await prisma.user.findUnique({
        where : {
            email,
        }
    })

    if(!isUserExist){
        throw new AppError(status.NOT_FOUND, "User not found");
    }

    if(!isUserExist.emailVerified){
        throw new AppError(status.BAD_REQUEST, "Email not verified");
    }

    if(isUserExist.isDeleted || isUserExist.status === UserStatus.DELETED){
        throw new AppError(status.NOT_FOUND, "User not found"); 
    }

    await auth.api.requestPasswordResetEmailOTP({
        body:{
            email,
        }
    })
}


const resetPassword = async (email : string, otp : string, newPassword : string) => {
    const isUserExist = await prisma.user.findUnique({
        where: {
            email,
        }
    })

    if (!isUserExist) {
        throw new AppError(status.NOT_FOUND, "User not found");
    }

    // if (!isUserExist.emailVerified) {
    //     throw new AppError(status.BAD_REQUEST, "Email not verified");
    // }

    if (isUserExist.isDeleted || isUserExist.status === UserStatus.DELETED) {
        throw new AppError(status.NOT_FOUND, "User not found");
    }

    await auth.api.resetPasswordEmailOTP({
        body:{
            email,
            otp,
            password : newPassword,
        }
    })

    if (isUserExist.needPasswordChange) {
        await prisma.user.update({
            where: {
                id: isUserExist.id,
            },
            data: {
                needPasswordChange: false,
            }
        })
    }

    await prisma.session.deleteMany({
        where:{
            userId : isUserExist.id,
        }
    })
}

const resendOTP = async (email: string) => {
  const user = await prisma.user.findUnique({
    where: { email },
  });

  if (!user) {
    throw new AppError(404, "User not found");
  }

  if (user.emailVerified) {
    throw new AppError(400, "Email already verified");
  }

  return await auth.api.sendVerificationEmail({
    body: {
      email,
    },
  });
};



export const AuthService = {
    loginUser,
    getNewToken,
    changePassword,
    logoutUser,
    verifyEmail,
    forgetPassword,
    resetPassword,
    resendOTP,
}